tests/wpt/tests/xhr/access-control-preflight-credential-sync.htm

<!DOCTYPE html>
<html>
  <head>
    <title>Tests proper handling of cross-origin sync request with credentials</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
  </head>
  <body>
    <script type="text/javascript">
    test(() => {
      const xhr = new XMLHttpRequest;

      xhr.open("PUT", get_host_info().HTTP_REMOTE_ORIGIN + "/xhr/resources/access-control-auth-basic.py?uid=fooUser", false, "fooUser", "barPass");

      xhr.withCredentials = true;

      xhr.send();

      assert_equals(xhr.status, 401, "Request raises HTTP 401: Unauthorized error.");
    }, "CORS sync request with URL credentials");
    </script>
  </body>
</html>
